mavishky dialog launch
git-svn-id: https://cgru.svn.sourceforge.net/svnroot/cgru@1178 0f86f178-d61e-0410-9d12-f6255c3815b8

+++ utilities/release/package/cgru.sh
@@ -33,4 +33,4 @@ cp -r $src/utilities/fixpathes.py $pack/$cgru/utilities
 cp -r $root/utilities/regexp/bin/regexp $pack/$cgru/utilities/regexp/bin
 cp $root/bin/exrjoin $pack/$cgru/bin/
 cp $root/bin/ffmpeg $pack/$cgru/bin/
-cp $root/bin/convert $pack/$cgru/bin/
+[ -f $root/bin/convert ] && cp $root/bin/convert $pack/$cgru/bin/
